feat(AuthService): refactor
