+++ p-week-01/TodoApp/Models/UserModel.cs
@@ -13,5 +13,6 @@ namespace TodoApp.Models
         public string Username { get; set; }
         [Required]
         public string Password { get; set; }
+        public string Token { get; set; }
     }
 }
